chore: 일련번호 데이터타입 long -> int 변환

+++ src/main/java/com/kosta/ems/student/StudentCourseInfoDTO.java
@@ -26,6 +26,6 @@ public class StudentCourseInfoDTO {
     private char gender;
     private String managerId;
     private String isActive;
-    private Long studentCourseSeq;
-    private Long courseSeq;
+    private int studentCourseSeq;
+    private int courseSeq;
 }
